+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [1.0.3] - 2026-05-24
+
+- fix(excludes): `!/regex/` exclude entries now strip the trailing `/`
+- refactor: forEach -> for..of
+- remove unneeded done callbacks
+- update triggers (#3)
+
 ### [1.0.2] - 2025-01-30
 
 - dep(eslint): upgrade to v9
@@ -21,3 +28,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [1.0.0]: https://github.com/haraka/haraka-plugin-clamd/releases/tag/v1.0.0
 [1.0.1]: https://github.com/haraka/haraka-plugin-clamd/releases/tag/v1.0.1
 [1.0.2]: https://github.com/haraka/haraka-plugin-clamd/releases/tag/v1.0.2
+[1.0.3]: https://github.com/haraka/haraka-plugin-clamd/releases/tag/v1.0.3